+++ lib/src/webview_helper/webview_helper_none.dart
@@ -1,0 +1,11 @@
+class WebViewHelper {
+  static List<String> get desktopBrowserAvailablePath => [];
+
+  static bool get isChromeAvailable => false;
+
+  static String? executablePath() => null;
+
+  static bool get isDesktop => false;
+
+  static bool get isMobile => false;
+}
