+++ pubspec.yaml
@@ -13,6 +13,7 @@ dependencies:
   puppeteer: ^2.11.0  
   dio: ^4.0.6  
   easy_logger: ^0.0.2
+  js: ^0.6.5
 
 dev_dependencies:
   flutter_test:
